+++ nbproject/private/private.xml
@@ -3,10 +3,8 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
+            <file>file:/D:/JavaSwing/Project%20Company%20Management/src/project/company/management/them_can_bo.java</file>
             <file>file:/D:/JavaSwing/Project%20Company%20Management/src/project/company/management/controller/ChuyenManHinhController.java</file>
-            <file>file:/D:/JavaSwing/Project%20Company%20Management/src/project/company/management/bean/DanhMucBean.java</file>
-            <file>file:/D:/JavaSwing/Project%20Company%20Management/src/project/company/management/dang_nhap3.java</file>
-            <file>file:/D:/JavaSwing/Project%20Company%20Management/src/project/company/management/view/Home.java</file>
             <file>file:/D:/JavaSwing/Project%20Company%20Management/src/project/company/management/view/MainFrame.java</file>
         </group>
     </open-files>
